Add check to see if custom theme is already in place

+++ installDotFiles.sh
@@ -56,13 +56,15 @@ cloneDockerProjects()
 }
 
 get_custom_zsh_theme(){
-  # Craig's theme is pretty rad
-  wget https://raw.githubusercontent.com/wunderhund/dotfiles/master/craig-magic.zsh-theme -O ~/.oh-my-zsh/themes/craig-magic.zsh-theme
+  if [ ! -f ~/.oh-my-zsh/themes/craig-magic.zsh-theme ]; then
+    # Craig's theme is pretty rad
+    wget https://raw.githubusercontent.com/wunderhund/dotfiles/master/craig-magic.zsh-theme -O ~/.oh-my-zsh/themes/craig-magic.zsh-theme
+  fi
 }
 
 # Backup old zshrc (if one exists)
 if [ -f ~/.zshrc ]; then
-    echo -e "${YELLOW}Backup up old zshrc, please wait...${RESET}"
+  echo -e "${YELLOW}Backup up old zshrc, please wait...${RESET}"
   mv $HOME/.zshrc $HOME/.zshrc.old
 fi
 
@@ -107,7 +109,7 @@ cp -r $installDir/files $dotdir/files
 
 # If janus is installed, move our .vimrc.after into place
 if [[ -d $HOME/.vim/janus ]]; then
-    cp $dotdir/files/.vimrc.after $HOME/.vimrc.after
+  cp $dotdir/files/.vimrc.after $HOME/.vimrc.after
 fi
 
 # Move .gitconfig into place
